requirements.txt
pyusb>=1.2.0
customtkinter>=5.2.0
